% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_date.R
\name{assign_date}
\alias{assign_date}
\title{assign_date()}
\usage{
assign_date(dates, new_date)
}
\arguments{
\item{dates}{character vector: vector of date characters}

\item{new_date}{character vector: Represent the date to be assigned to missing values}
}
\value{
vector
}
\description{
Helper function. Assign date to missing records
}
\examples{
assign_date(my_dates, new_date = '01/01/2015')
}

