% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_activity.R
\name{check_activity}
\alias{check_activity}
\title{check_activity()}
\usage{
check_activity(enroll_file, pos_file, tp_file, asmt_file, ref_file,
  save_output = FALSE)
}
\arguments{
\item{enroll_file}{character vector: the name of the file which the enrollment data are to be read from. This must be a .txt file exported from the [Admin] raw_enrollment_report}

\item{pos_file}{character vector: the name of the file which the "point of service" data are to be read from. This must be a .txt file exported from the `[Admin] raw_pos_report`}

\item{tp_file}{character vector: the name of the file which the "touchpoint" data are to be read from. This must be a .txt file exported from the `[Admin] raw_touchpoint_report`}

\item{asmt_file}{character vector: the name of the file which the "assessment" data are to be read from. This must be a .txt file exported from the `[Admin] raw_assessment_report`}

\item{ref_file}{character vector: the name of the file which the "referral" data are to be read from. This must be a .txt file exported from the `[Admin] raw_referral_report}

\item{save_output}{logical: FALSE: Return output on screen, TRUE: Save output in the "output" folder in the working directory}
}
\value{
plot and data frame
}
\description{
This function checks if active ETO partipants have some service activity recorded through either "point of service", "touchpoint", "assessment" or "referrals".
}
\examples{
check_activity(enroll_file = './data/enroll.txt', pos_file = './data/pos.txt', asmt_file = './data/asmt.txt', ref_file = './data/ref.txt')
}

