% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{clean_data()}
\usage{
clean_data(data, to_date = c("start", "end", "date"),
  to_numeric = c("days_enrolled", "time"))
}
\arguments{
\item{data}{dataframe: a dataframe containing.}

\item{to_date}{character vector: Names of columns to be converted to date format.}

\item{to_numeric}{character vector: Names of columns to be converted to numeric format.}
}
\value{
dataframe
}
\description{
This function conducts basic data cleaning tasks on a data frame.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
clean_data(data = enroll)
}

