% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_id}
\alias{create_id}
\title{create_id()}
\usage{
create_id(df, var = c("first_name", "last_name"))
}
\arguments{
\item{df}{dataframe}

\item{var}{character vector: list of variables to be concatenated to form the new id}
}
\value{
character vector
}
\description{
Helper function. Create custom id by concatenating different variables
}
\examples{
create_id(df, c('first_name', 'last_name')
}

