% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{format_data()}
\usage{
format_data(data, to_date = c("start$|^start|end$|^end|date"),
  to_numeric = c("id$|days|time|weight|wage"))
}
\arguments{
\item{data}{dataframe: a dataframe containing.}

\item{to_date}{character vector: Names of columns to be converted to date format.}

\item{to_numeric}{character vector: Names of columns to be converted to numeric format.}
}
\value{
dataframe
}
\description{
This function conducts basic data cleaning tasks on a data frame.
}
\examples{
enroll <- laycUtils::load_txt('./my_data_folder/enrollment.txt')
format_data(data = enroll)
}

