% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_headers.R
\name{format_headers}
\alias{format_headers}
\title{format_headers()}
\usage{
format_headers(df)
}
\arguments{
\item{data}{dataframe: a dataframe}
}
\value{
dataframe
}
\description{
This function conducts format headers so all variable names are lower case, and separated by an underscore
}
\examples{
my_dataframe <- laycUtils::format_headers(my_dataframe)
}

