% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzy_join.R
\name{fuzzy_join}
\alias{fuzzy_join}
\title{fuzzy_join()}
\usage{
fuzzy_join(x, y, by)
}
\arguments{
\item{x,y}{dataframe to merge}

\item{by}{a character variable to join by. This variable is generally the otput of the create_id() function}
}
\value{
dataframe
}
\description{
Merge 2 dataframes with inconsistent ID variables (due to typos for instance).
}
\examples{
fuzzy_join(x = dataframe1, y = dataframe2, by = 'custom_id')
}

