% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/id_prepost.R
\name{id_prepost}
\alias{id_prepost}
\title{id_prepost()}
\usage{
id_prepost(df, date)
}
\arguments{
\item{df}{dataframe: a dataframe containing longitudinal data.}
}
\value{
dataframe
}
\description{
This is a helper function that identifies 'pre' and 'post' test based on the date the test was taken.
}
\examples{
df <- id_prepost(df, date = 'test_date')
}

