% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change.R
\name{plot_change}
\alias{plot_change}
\title{plot_change()}
\usage{
plot_change(var, my_palette = c("#2CA02C", "#FF7F0E", "#D62728"))
}
\arguments{
\item{var:}{The variable to plot}
}
\value{
dataframe
}
\description{
Takes a continuous variable with values from -Inf to +Inf, recodes the variable into three categories: negative, no change, positive, and plot the results.
}

