% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_layc.R
\name{theme_layc}
\alias{theme_layc}
\title{Theme for LAYC plots}
\usage{
theme_layc(base_size = 12, base_family = "Calibri")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Inspired by theme_few() from ggthemes package
}
\examples{
plot <- ggplot(diamonds, aes(x = , y = )) + geom_point()
layc_plot <- plot + theme_layc
}
\references{
ggthemes
\url{https://github.com/jrnold/ggthemes}.
}

